Move lurk_macro into lurk-derive.
