+++ Cargo.toml
@@ -18,7 +18,7 @@ ff = "0.12.1"
 generic-array = "0.14.6"
 itertools = "0.9"
 log = "0.4.17"
-lurk-derive = { path = "lurk-derive" }
+lurk-macros = { path = "lurk-macros" }
 neptune = { version = "8.1.1", features = ["arity2","arity4","arity8","arity16","pasta","bls"] }
 nova = { package = "nova-snark", version = "0.19.0", default-features = false }
 once_cell = "1.17.1"
@@ -76,7 +76,7 @@ harness = false
 
 [workspace]
 members = [
-  "lurk_macro", "fcomm", "clutch", "lurk-derive"
+ "fcomm", "clutch", "lurk-macros"
 ]
 
 [[bin]]
