+++ lurk-macros/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "lurk-derive"
+name = "lurk-macros"
 version = "0.1.0"
 authors = ["porcuquine <porcuquine@gmail.com>"]
 license = "MIT OR Apache-2.0"
@@ -11,7 +11,10 @@ repository = "https://github.com/lurk-lab/lurk-rs"
 proc-macro = true
 
 [dependencies]
-syn = { version = "1.0.64", features = ["derive", "full"] }
+syn = { version = "1.0.64", features = ["derive", "extra-traits", "full"] }
 quote = "1.0.9"
 proc-macro2 = "1.0.24"
 
+[dev-dependencies]
+lurk_crate = { path = "../", package = "lurk" }
+pasta_curves = { version = "0.5.2", features = ["repr-c", "serde"], package = "fil_pasta_curves" }
