--- lurk_macro/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "lurk_macro"
-version = "0.1.0"
-authors = ["porcuquine <porcuquine@gmail.com>"]
-license = "MIT OR Apache-2.0"
-description = "Lurk in Rust"
-edition = "2021"
-
-[dependencies]
-blstrs = "0.6.1"
-lurk = { path = "../", package = "lurk" }
-proc-macro2 = "1"
-quote = "1"
-syn = { version = "1", features = ["extra-traits", "full"] }
-
-[lib]
-proc-macro = true
